+++ src/Null/Null.hpp
@@ -14,6 +14,7 @@ struct BlankModule : ResizableModule
     bool glow = false;
     bool branding = true;
     bool copper = true;
+    bool brand_logo = false;
     bool dirty_settings = false;
 
     bool is_bright() { return bright; }
@@ -24,6 +25,8 @@ struct BlankModule : ResizableModule
 
     bool is_branding() { return branding; }
     void set_branding(bool b) { branding = b; }
+    bool is_brand_logo() { return brand_logo; }
+    void set_brand_logo(bool b) { brand_logo = b; }
 
     BlankModule();
     void onReset(const ResetEvent& e) override;
@@ -79,6 +82,9 @@ struct BlankModuleWidget : ModuleWidget, IThemeChange
     }
     bool branding() {
         return my_module ? my_module->is_branding() : true;
+    }
+    bool brand_logo () {
+        return my_module ? my_module->is_brand_logo() : true;
     }
 
     explicit BlankModuleWidget(BlankModule* module);